4_3RAG/requirements.txt
# RAG System Requirements
# Had to install these one by one when stuff wasn't working

beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.1.0
sentence-transformers==2.2.2
numpy==1.24.3
faiss-cpu==1.7.4
transformers==4.36.2
torch==2.1.2

# Also tried these but didn't end up needing them:
# PyPDF2==3.0.1
# pdfminer.six==20221105

# Note: torch is huge (like 2GB) so download takes forever
# sentence-transformers downloads models on first use - also slow
